lidar/requirements.txt
pyserial>=3.5
matplotlib>=3.7.0
numpy>=1.23
